requirements.txt
pandas
requests
transformers
pytorch
dall_e
numpy